+++ src/Octoshift/Models/BbsRepository.cs
@@ -1,0 +1,8 @@
+namespace Octoshift.Models;
+
+public record BbsRepository
+{
+    public string Id { get; init; }
+    public string Name { get; init; }
+    public string Slug { get; init; }
+}